관리자 개인리포트 목록 검색 api 추가

+++ src/main/java/kr/co/talk/domain/statistics/dto/AdminReportListDto.java
@@ -14,5 +14,5 @@ public class AdminReportListDto {
     private String profileUrl;
     private String name;
     private String nickname;
-    private int chatCount;
+    private long chatCount;
 }
